Also load pfam domain colors for local files
